requirements.txt
numpy==1.19.5
Keras==2.4.3
keras-contrib==2.0.8
Keras-Preprocessing==1.1.2
keras-vis==0.4.1
matplotlib==3.2.2
pandas==1.1.5
Pillow==7.1.2
scikit-image==0.16.2
scikit-learn==0.22.2.post1
scipy==1.4.1
tensorflow==2.4.1
